Ingredients table removed.
Instead of relating ingredients to a recipe with an additional table 'Ingredients', the table 'Recipe' will hold all ingredients in a string which will be parsed such that '@@@@@' will be used to separate different ingredients from each other. This might not be a good idea because if an ingredient has '@@@@@' it could break the parsing function. Might have to come up with a better solution, but this works for now.